backend/requirements.txt
fastapi
uvicorn[standard]
motor
python-jose[cryptography]
passlib[bcrypt]
langchain
langchain-community
langchain-google-genai
google-generativeai
chromadb
python-multipart
python-dotenv
pydantic[email]
pydantic-settings
